Removed duplicated BDD tests.
