+++ .vortex/installer/composer.json
@@ -18,7 +18,7 @@
     },
     "require": {
         "php": ">=8.2",
-        "alexskrypnyk/file": "^0.10",
+        "alexskrypnyk/file": "^0.11",
         "alexskrypnyk/str2name": "^1.4",
         "composer/composer": "^2.8",
         "cweagans/composer-patches": "^1.7",
