+++ .vortex/tests/phpunit/Functional/FunctionalTestCase.php
@@ -8,22 +8,29 @@ use AlexSkrypnyk\File\File;
 use AlexSkrypnyk\PhpunitHelpers\Traits\AssertArrayTrait;
 use AlexSkrypnyk\PhpunitHelpers\Traits\EnvTrait;
 use AlexSkrypnyk\PhpunitHelpers\Traits\LocationsTrait;
+use AlexSkrypnyk\PhpunitHelpers\Traits\LoggerTrait;
 use AlexSkrypnyk\PhpunitHelpers\Traits\ProcessTrait;
 use AlexSkrypnyk\PhpunitHelpers\UnitTestCase;
 use DrevOps\Vortex\Tests\Traits\AssertProjectFilesTrait;
 use DrevOps\Vortex\Tests\Traits\GitTrait;
-use AlexSkrypnyk\PhpunitHelpers\Traits\LoggerTrait;
 use DrevOps\Vortex\Tests\Traits\Steps\StepBuildTrait;
-use DrevOps\Vortex\Tests\Traits\Steps\StepDownloadDbTrait;
+use DrevOps\Vortex\Tests\Traits\Steps\StepDatabaseTrait;
 use DrevOps\Vortex\Tests\Traits\Steps\StepPrepareSutTrait;
-use DrevOps\Vortex\Tests\Traits\Steps\StepTestBddAllTrait;
-use DrevOps\Vortex\Tests\Traits\Steps\StepTestBddTrait;
+use DrevOps\Vortex\Tests\Traits\Steps\StepTestTrait;
 use Symfony\Component\Process\Process;
 
 /**
  * Base class for functional tests.
  */
 class FunctionalTestCase extends UnitTestCase {
+
+  /**
+   * URL to the test demo database.
+   *
+   * Tests use demo database and 'ahoy download-db' command, so we need
+   * to set the CURL DB to test DB.
+   */
+  const VORTEX_INSTALLER_DEMO_DB_TEST = 'https://github.com/drevops/vortex/releases/download/25.4.0/db_d11_2.test.sql';
 
   use AssertArrayTrait;
   use AssertProjectFilesTrait;
@@ -35,10 +42,9 @@ class FunctionalTestCase extends UnitTestCase {
     ProcessTrait::processRun as traitProcessRun;
   }
   use StepBuildTrait;
-  use StepDownloadDbTrait;
   use StepPrepareSutTrait;
-  use StepTestBddAllTrait;
-  use StepTestBddTrait;
+  use StepTestTrait;
+  use StepDatabaseTrait;
 
   protected function setUp(): void {
     self::locationsInit(File::cwd() . DIRECTORY_SEPARATOR . '..' . DIRECTORY_SEPARATOR . '..');
@@ -163,6 +169,72 @@ class FunctionalTestCase extends UnitTestCase {
 
   public function volumesMounted(): bool {
     return getenv('VORTEX_DEV_VOLUMES_SKIP_MOUNT') != 1;
+  }
+
+  protected function trimFile(string $file): void {
+    $content = File::read($file);
+    $lines = explode("\n", $content);
+    // Remove last line.
+    array_pop($lines);
+    File::dump($file, implode("\n", $lines));
+  }
+
+  protected function stepWarmCaches(): void {
+    $this->logSubstep('Warming up caches');
+    $this->cmd('ahoy drush cr');
+    $this->cmd('ahoy cli curl -- -sSL -o /dev/null -w "%{http_code}" http://nginx:8080 | grep -q 200');
+  }
+
+  protected function addVarToFile(string $file, string $var, string $value): void {
+    // Backup original file first.
+    $this->backupFile($file);
+    $content = File::read($file);
+    $content .= sprintf('%s%s=%s%s', PHP_EOL, $var, $value, PHP_EOL);
+    File::dump($file, $content);
+  }
+
+  protected function backupFile(string $file): void {
+    $backup_dir = '/tmp/bkp';
+    if (!is_dir($backup_dir)) {
+      mkdir($backup_dir, 0755, TRUE);
+    }
+    File::copy($file, $backup_dir . '/' . basename($file));
+  }
+
+  protected function restoreFile(string $file): void {
+    $backup_file = '/tmp/bkp/' . basename($file);
+    if (file_exists($backup_file)) {
+      File::copy($backup_file, $file);
+    }
+  }
+
+  protected function createDevelopmentSettings(string $webroot = 'web'): void {
+    File::copy($webroot . '/sites/default/example.settings.local.php', $webroot . '/sites/default/settings.local.php');
+    // Assert manually created local settings file exists.
+    $this->assertFileExists($webroot . '/sites/default/settings.local.php');
+
+    File::copy($webroot . '/sites/default/example.services.local.yml', $webroot . '/sites/default/services.local.yml');
+    // Assert manually created local services file exists.
+    $this->assertFileExists($webroot . '/sites/default/services.local.yml');
+  }
+
+  protected function removeDevelopmentSettings(string $webroot = 'web'): void {
+    File::remove([
+      $webroot . '/sites/default/settings.local.php',
+      $webroot . '/sites/default/services.local.yml',
+    ]);
+    $this->assertFileDoesNotExist($webroot . '/sites/default/settings.local.php');
+    $this->assertFileDoesNotExist($webroot . '/sites/default/services.local.yml');
+  }
+
+  protected function assertFilesPresent(string $webroot): void {
+    // Use existing method from base class with correct signature.
+    $this->assertCommonFilesPresent($webroot);
+  }
+
+  protected function assertGitRepo(): void {
+    // @todp Use gitAssertIsRepository().
+    $this->assertDirectoryExists('.git');
   }
 
 }
