+++ .vortex/tests/phpunit/Traits/Steps/StepDatabaseTrait.php
@@ -11,6 +11,22 @@ use AlexSkrypnyk\File\File;
  */
 trait StepDatabaseTrait {
 
+  protected function stepDownloadDb(): void {
+    $this->logStepStart();
+
+    File::remove('.data/db.sql');
+    $this->assertFileDoesNotExist('.data/db.sql', 'File .data/db.sql should not exist before downloading the database.');
+
+    $this->logSubstep('Downloading demo database from ' . static::VORTEX_INSTALLER_DEMO_DB_TEST);
+    $this->cmd('ahoy download-db', env: [
+      'VORTEX_DB_DOWNLOAD_URL' => static::VORTEX_INSTALLER_DEMO_DB_TEST,
+    ]);
+
+    $this->assertFileExists('.data/db.sql', 'File .data/db.sql should exist after downloading the database.');
+
+    $this->logStepFinish();
+  }
+
   protected function stepAhoyExportDb(string $filename = ''): void {
     $this->logStepStart();
