+++ .vortex/tests/phpunit/Traits/Steps/StepEnvironmentTrait.php
@@ -202,55 +202,4 @@ trait StepEnvironmentTrait {
     $this->logStepFinish();
   }
 
-  protected function addVarToFile(string $file, string $var, string $value): void {
-    // Backup original file first.
-    $this->backupFile($file);
-    $content = File::read($file);
-    $content .= sprintf('%s%s=%s%s', PHP_EOL, $var, $value, PHP_EOL);
-    File::dump($file, $content);
-  }
-
-  protected function backupFile(string $file): void {
-    $backup_dir = '/tmp/bkp';
-    if (!is_dir($backup_dir)) {
-      mkdir($backup_dir, 0755, TRUE);
-    }
-    File::copy($file, $backup_dir . '/' . basename($file));
-  }
-
-  protected function restoreFile(string $file): void {
-    $backup_file = '/tmp/bkp/' . basename($file);
-    if (file_exists($backup_file)) {
-      File::copy($backup_file, $file);
-    }
-  }
-
-  protected function createDevelopmentSettings(string $webroot = 'web'): void {
-    File::copy($webroot . '/sites/default/example.settings.local.php', $webroot . '/sites/default/settings.local.php');
-    // Assert manually created local settings file exists.
-    $this->assertFileExists($webroot . '/sites/default/settings.local.php');
-
-    File::copy($webroot . '/sites/default/example.services.local.yml', $webroot . '/sites/default/services.local.yml');
-    // Assert manually created local services file exists.
-    $this->assertFileExists($webroot . '/sites/default/services.local.yml');
-  }
-
-  protected function removeDevelopmentSettings(string $webroot = 'web'): void {
-    File::remove([
-      $webroot . '/sites/default/settings.local.php',
-      $webroot . '/sites/default/services.local.yml',
-    ]);
-    $this->assertFileDoesNotExist($webroot . '/sites/default/settings.local.php');
-    $this->assertFileDoesNotExist($webroot . '/sites/default/services.local.yml');
-  }
-
-  protected function assertFilesPresent(string $webroot): void {
-    // Use existing method from base class with correct signature.
-    $this->assertCommonFilesPresent($webroot);
-  }
-
-  protected function assertGitRepo(): void {
-    $this->assertDirectoryExists('.git');
-  }
-
 }
